+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## Update 0.5.2
 
 - Skyblock ProfileId is now parsed from chat.
+- Added Cursor saving between pages changes in the overlay